Make slider dots more visible using sticky position
Used only on small and medium sized screens: the slider dots will now stick to the bottom of the screen when the slider container element is in view (and the bottom of the slider container element is not). As soon the bottom of the slider element is in view, the dots will show as normal right below the slides.